requirements.txt
pyyaml
stata_setup
pystata
pandas
numpy